requirements.txt
-e .
aws-cdk.core==1.145.0
aws-cdk.aws_apigateway==1.145.0
aws-cdk.aws_certificatemanager==1.145.0
aws-cdk.aws_cloudformation==1.145.0
aws-cdk.aws_cloudfront==1.145.0
aws-cdk.aws_cloudfront_origins==1.145.0
aws-cdk.aws_dynamodb==1.145.0
aws-cdk.aws_iam==1.145.0
aws-cdk.aws_iot==1.145.0
aws-cdk.aws_lambda==1.145.0
aws-cdk.aws_route53==1.145.0
aws-cdk.aws_route53_patterns==1.145.0
aws-cdk.aws_route53_targets==1.145.0
aws-cdk.aws_s3==1.145.0
aws-cdk.aws_s3_deployment==1.145.0
aws-cdk.aws_secretsmanager==1.145.0
aws-cdk.aws_ses==1.145.0
aws-cdk.aws_ssm==1.145.0
aws-cdk.aws_wafv2==1.145.0
boto3==1.18.50
pylint==2.8.3
pytest==6.2.4
pytest-dependency==0.5.1
requests==2.25.1
prance[osv,cli]==0.21.2
yapf==0.31.0

